requirements.txt
Flask==3.1.0
python-dotenv
psycopg2-binary